EntityPresenter, RepositoryPresenter and pattern matching added
This update adds the ability to have one presenter to multiple views, by pattern matching the posible executing mechanisms that can be implemented by a Repository. A Repository can have a complete set of methods for the CRUD operations, depending on the contracts that it implements. All this allows developers to switch from "default repository behaviors" to custom behaviors depending on how the management of an entity evolves in time. Presenters are now responsible for sending entity events through the MessageCenter from WebPx.Events library. This events can then be translated to user friendly notifications as Status messages, sent throught the Status.Add method.